Custom brew cask file for Script 12 BT. Corrected issues in install script.

+++ install.sh
@@ -10,31 +10,45 @@ sh -c "$(curl -fsSL https://raw.githubusercontent.com/robbyrussell/oh-my-zsh/mas
 /usr/bin/ruby -e "$(curl -fsSL https://raw.githubusercontent.com/Homebrew/install/master/install)"
 
 # personal settings
-git clone https://github.com/jerrythomas/personalize.git
-cd personalize
+#git clone https://github.com/jerrythomas/personalize.git
+#cd personalize
 
 # zsh theme
+pwd
+
+if [ -f ~/.vimrc ]; then
+  mv ~/.vimrc ~/.vimrc.old
+fi
+cp themes/template.vimrc ~/.vimrc
+
+if [ ! -d ~/.atom ]; then
+  mkdir -p ~/.atom
+fi
+
 cp themes/clarity.zsh-themes ~/.oh-my-zsh/themes/.
-mv ~/.zshrc ~/.zshrc.old
+if [ -f ~/.atom/styles.less ]; then
+  mv ~/.atom/styles.less ~/.atom/styles.old
+fi
+cp themes/styles.less ~/.atom
+
+if [ -f ~/.zshrc ]; then
+   mv ~/.zshrc ~/.zshrc.old
+fi
 sed 's/robbyrussel/clarity/' ~/.zshrc.old > ~/.zshrc
 
 # terminal colors
-cp themes/template.vimrc ~/.vimrc
 
 # Fonts
 brew tap caskroom/fonts
 brew cask install font-fira-code
 
 # Cursive font Script 12 BT
-wget https://www.wfonts.com/download/data/2016/07/14/script12-bt/script12-bt.zip
-unzip script12-bt.zip
-sudo mv script12-bt/*.ttf ~/Library/Fonts
-rm -rf script12-bt script12-bt.zip
+brew cask install ./font-script12-bt.rb
 
+chmod 700 brew.zsh
+chmod 700 node.zsh
 # Brew apps
-brew.zsh
+./brew.zsh
 # Node libraries
-node.zsh
+./node.zsh
 # Python3
-
-cp themes/styles.less ~/.atom/
